JavaScript is available in str, dispFile is now automatic
First off, this is backwards compatible and works for Python 2 & 3.
The difference is that calling chart methods does not automatically output them
to file anymore. There's a new write() method for that, and wb() or nb() for
choosing the viewing method. The nb parameter to the chart methods can be used
to tell gpcharts to output after creating the chart. the chart_nb methods use
this to work in the same way they always have.

I added storage of the JavaScript within the object and the ability to get that
text via str(). The reason is that if you want more than one figure on a page or
to include the JS in another Python project, then you couldn't do that with
gpcharts. Now you can.

The boolean parameter to dispFile now does nothing. The method now automatically
chooses between notebook and web browser when called. A user can now
deliberately use nb() and wb() to choose as well. The bigger reason for this
change was that, now that the JS output is saved for possible printing, it
doesn't make sense to automatically output it to file anymore. You still have
that option, but it's not automatic.