Added get_drawChart, fixed quiet bug
There were two bugs in _template.py, both of which broke the end product.
Firstly, I forgot the s flag on a format string, so that broke. Secondly,
I forgot to add the figure name to the actual calling of drawChart_XXX(), so it
wasn't being called and the page was blank. Everything works now.